Add files via upload
Changed a few of the details for generating the model, to further simplify the input. Specifically, this pertains to the spin dictionary, which now includes not only spin-orbit coupling, but also AFM, FM ordering. In addition, you can now pass slab dictionary to the tight-binding model generation. Please see FeSe.py for a sample of how this is to be done.