Important Fix for Rotated SOC and spin
Previous iterations were not correctly implementing the rotation of spin-degrees of freedom (currently restricted to z-direction rotations) in the subsequent calculation of Spin-Orbit Coupling. Have introduced a simple fix, including independent 'projection' matrices for the cubic to spherical harmonics, which has been tested and confirmed for the case of Bi2Se3, where now a true chiral spin-texture is observed. Please update when possible.