only Graham Schmidt when necessary!
fixed some limits on filling in the remaining orbitals for orbital bases which use only a shell-subset: the projection matrix is filled with dummy vectors based on conventional cubic harmonic decompositions. These are then orthogonalized, but only as strictly necessary--some strange behaviour comes about when trying to orthogonalize against things which were already to floating point precision so. 

This allows now for effective SOC calculations in the presence of distortions (at least distortions around the z-axis for now***)