Low-energy Hamiltonian model of FeSe
This script is a first attempt at implementing low-energy effective models into the TB_ARPES framework. A bit awkward, but a small modification to TB_lib provides a route to defining an alternative Hamiltonian function which does not require the TB.mat_els. c.f. update to ARPES_lib for details.