handling effective Hamiltonians
We can now pass Hamiltonians which are not tight-binding models. Rather than pass a list of TB hopping elements, in the absence of a TB.mat_els list, TB.solve_H() looks for the presence of a TB.alt_ham()-- a function of momentum only which returns a Hamiltonian matrix in the same way as usually done. This allows then for definition of an effective Hamiltonian model, rather than relying on tight-binding alone.