intensity_maps
created an intensity map class for holding the arpes calculations in a self-contained object which has not only the intensity, but also the relevant experimental parameters. This allows for convenient connection to the tkinter gui, where these attributes are connected to a specific map, eliminating the need to have a parameter dictionary passed to all these functions.