fixed bug in O_path
O_path was ineffective in averaging over the top-degenerate bands. the control flow statements were not catching the case where we were at the last band -- in which case there would be no next state to catch a set of degeneracies. this has been fixed by changing an elif to an independent if statement. this was a problem since: if the last band was degenerate with the prior, then the elif would never be queried since the initial if was satisfied. by changing the elif |E-Elast|>tol or *last_band*, to an if "", this is solved. now O_path with degen=True runs nicely over all states.